ENH: removing usage of fiducial/markup, seedNode now uses double type, fixing tests based on changes
Refactoring logic so it no longer uses fiducial or markup, but instead seedNodes.
Removing fiducial/markup stuff from Displayable Manager and cleaning it up
Removing fiducial/markup stuff from Module Widget
Changing the SeedNode to use Double instead of Float
Fixes seed node test based on change to double
Tried to fix the segmentation test (but it still does not segment as far as the similarity image, so it fails the test)

+++ VesselSegmentation/MRML/vtkMRMLVesselSegmentationSeedNode.h
@@ -94,11 +94,11 @@ public:
    */
   virtual const char* GetIcon() {return "";}
 
-  vtkSetVector3Macro(Seed1, float);
-  vtkGetVector3Macro(Seed1, float);
+  vtkSetVector3Macro(Seed1, double);
+  vtkGetVector3Macro(Seed1, double);
 
-  vtkSetVector3Macro(Seed2, float);
-  vtkGetVector3Macro(Seed2, float);
+  vtkSetVector3Macro(Seed2, double);
+  vtkGetVector3Macro(Seed2, double);
 
 protected:
   vtkMRMLVesselSegmentationSeedNode();
@@ -107,8 +107,8 @@ protected:
   vtkMRMLVesselSegmentationSeedNode(const vtkMRMLVesselSegmentationSeedNode&);
   void operator=(const vtkMRMLVesselSegmentationSeedNode&);
 
-  float Seed1 [3];
-  float Seed2 [3];
+  double Seed1 [3];
+  double Seed2 [3];
 };
